pyproject.toml
[tool.poetry]
name = "domjudge-utility"
version = "0.0.4"
description = "DOMjudge Utility"
authors = ["Dup4 <lyuzhi.pan@gmail.com>"]
license = "MIT"
readme = "README.md"
classifiers = [
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"
xlwt = "^1.3.0"
requests = "^2.31.0"
requests-toolbelt = "^1.0.0"
grequests = "^0.6.0"
bs4 = "^0.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
flake8 = "^6.0.0"
pyright = "^1.1.305"
autopep8 = "^2.0.2"
pytest-snapshot = "^0.9.0"
poetry-bumpversion = "^0.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry_bumpversion.replacements]]
files = ["domjudge_utility/__init__.py"]
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'
